review related suggestions Related to Issue [WTI-1250]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@9429 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 31945adcc497c91c593e1911e2f97491eb44ae27
Former-commit-id: 7de59d7247c42e0c572fa25796134a00ea133286

+++ Application/Ringtoets/src/Application.Ringtoets.Migration/RingtoetsProjectMigrator.cs
@@ -147,13 +147,13 @@ namespace Application.Ringtoets.Migration
 
         private bool MigrateToTargetLocation(string sourceFilePath, string targetLocation)
         {
+            if (!CreateInitializedDatabaseLogFile())
+            {
+                return false;
+            }
+
             try
             {
-                if (!CreateInitializedDatabaseLogFile())
-                {
-                    return false;
-                }
-
                 var versionedFile = new RingtoetsVersionedFile(sourceFilePath);
                 fileMigrator.Migrate(versionedFile, currentDatabaseVersion, targetLocation);
                 string message = string.Format(Resources.RingtoetsProjectMigrator_MigrateToTargetLocation_Outdated_projectfile_0_succesfully_updated_to_target_filepath_1_version_2_,
@@ -216,7 +216,10 @@ namespace Application.Ringtoets.Migration
         {
             try
             {
-                File.Delete(migrationLogPath);
+                if (File.Exists(migrationLogPath))
+                {
+                    File.Delete(migrationLogPath);
+                }
             }
             catch (SystemException exception) when (exception is IOException
                                                     || exception is UnauthorizedAccessException)
